Extract a repo_state function
Rather than using an inline ternary to work out the class to attach to
the `tr`, extract the conditional to a `repo_state` method and pass it
the `repo` object.

This change will make it simpler to vary the class we attach based on
other factors, like how long the repo has been stale for.